consolidate appdata + appstate
